fix: remove react query chain to fetch results
